+++ CHANGELOG.md
@@ -74,6 +74,7 @@
     - Bug-084686 - User is still able to activate the disabled field from console. 
     - Bug-085023 - Global RFQ flow is not working properly
     - Bug-085117 - Address Label Field Exceeds Character Limit in Shipping Address Form
+    - Bug-085115 - 'Duplicate offer with same quantity' error still appears when admin tries to add an offer for the same seller after deleting the previous offer.
 
 ## Hit URL:
     - admin/patch-update/updateProductRating